test: Add CodeGenTest
