+++ core/src/test/java/com/tlcsdm/core/freemarker/CodeGenTest.java
@@ -27,12 +27,20 @@
 
 package com.tlcsdm.core.freemarker;
 
+import com.tlcsdm.core.exception.CoreException;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Test;
 
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.nio.file.Files;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
 /**
  * @author unknowIfGuestInDream
@@ -50,6 +58,44 @@ class CodeGenTest {
         }
         installDir.mkdirs();
         System.out.println(Files.createTempDirectory(installDirName));
+    }
+
+    private void unzip(String zipPath, String destDirPath) throws IOException, CoreException {
+        byte[] buf = new byte[8192];
+        File destDir = new File(destDirPath);
+        try (InputStream zipIn = new FileInputStream(zipPath); ZipInputStream zis = new ZipInputStream(
+            new BufferedInputStream(zipIn))) {
+            ZipEntry zEntry;
+            while ((zEntry = zis.getNextEntry()) != null) {
+                // if it is empty directory, create it
+                if (zEntry.isDirectory()) {
+                    new File(destDir, zEntry.getName()).mkdirs();
+                    continue;
+                }
+                // if it is a file, extract it
+                String filePath = zEntry.getName();
+                int lastSeparator = filePath.lastIndexOf("/");
+                String fileDir = "";
+                if (lastSeparator >= 0) {
+                    fileDir = filePath.substring(0, lastSeparator);
+                }
+                // create directory for a file
+                new File(destDir, fileDir).mkdirs();
+                // write file
+                String destDirCanonicalPath = destDir.getCanonicalPath();
+                File outFile = new File(destDir, filePath);
+                String outFileCanonicalPath = outFile.getCanonicalPath();
+                if (!outFileCanonicalPath.startsWith(destDirCanonicalPath + File.separator)) {
+                    throw new CoreException("Entry is outside of the target dir: " + filePath);
+                }
+                try (BufferedOutputStream outputStream = new BufferedOutputStream(new FileOutputStream(outFile))) {
+                    int n;
+                    while ((n = zis.read(buf)) >= 0) {
+                        outputStream.write(buf, 0, n);
+                    }
+                }
+            }
+        }
     }
 
 }
